Experimenting with Adding Support for Schemes in a Different Way
I need to talk with Martin again to see exactly what he was trying to
accomplish with his support for schemes. At any rate, his method was
not going to work for our needs if we are going to start using
CocoaPods. His method tried to find the workspace and change the
project path based on that finding if there was a workspace, then he
tried to inject some code into the xcscheme file. This is problematic
as the xcscheme file is generated for each user and is subject to
change (and we typically do not commit these).