Adding support for specifying action type for xcodebuild command
Looks for <action /> in the plugin's configuration block. If it's not
there, defaults to build. If it is there, it validates that it's either
build or archive. If it's not one of those, it throws an
IllegalStateException.

I also removed my experimental workspaceScheme stuff.

I also fast forwarded to the latest from hardison.

Updated the README to reflect new <action /> configuration variable.