Fixing defect where if workspace was being used, the application would die when trying to load a shared scheme
Always tried to look in the .xcodeproj file for this; In a workspace
setup, this will be in the .xcworkspace file. I have made a change so
that it will pull from the correct area. Either way, this part is now
skipped. It looks like we keep pulling from the project file even if we
are specifying workspace.